+++ src/graph/query_builder.rs
@@ -168,17 +168,21 @@ impl<Out, T: Display> QueryBuilder<'_, Out, T, SyncGraph> {
         let mut params = vec![query.as_str(), "--compact"];
         params.extend(timeout.as_deref());
 
-        self.graph
-            .get_client()
-            .borrow_connection(self.graph.get_client().clone())
-            .and_then(|mut conn| {
-                conn.execute_command(
-                    Some(self.graph.graph_name()),
-                    self.command,
-                    None,
-                    Some(params.as_slice()),
-                )
-            })
+        let client = self.graph.get_client();
+        let conn = if self.command == "GRAPH.RO_QUERY" {
+            client.borrow_readonly_connection(client.clone())
+        } else {
+            client.borrow_connection(client.clone())
+        };
+
+        conn.and_then(|mut conn| {
+            conn.execute_command(
+                Some(self.graph.graph_name()),
+                self.command,
+                None,
+                Some(params.as_slice()),
+            )
+        })
     }
 }
 
@@ -195,10 +199,14 @@ impl<'a, Out, T: Display> QueryBuilder<'a, Out, T, AsyncGraph> {
         let mut params = vec![query.as_str(), "--compact"];
         params.extend(timeout.as_deref());
 
-        self.graph
-            .get_client()
-            .borrow_connection(self.graph.get_client().clone())
-            .await?
+        let client = self.graph.get_client();
+        let conn = if self.command == "GRAPH.RO_QUERY" {
+            client.borrow_readonly_connection(client.clone()).await
+        } else {
+            client.borrow_connection(client.clone()).await
+        };
+
+        conn?
             .execute_command(
                 Some(self.graph.graph_name()),
                 self.command,
@@ -406,17 +414,21 @@ impl<Out> ProcedureQueryBuilder<'_, Out, SyncGraph> {
             generate_procedure_call(self.procedure_name, self.args, self.yields);
         let query = construct_query(query_string, params.as_ref());
 
-        self.graph
-            .get_client()
-            .borrow_connection(self.graph.get_client().clone())
-            .and_then(|mut conn| {
-                conn.execute_command(
-                    Some(self.graph.graph_name()),
-                    command,
-                    None,
-                    Some(&[query.as_str(), "--compact"]),
-                )
-            })
+        let client = self.graph.get_client();
+        let conn = if self.readonly {
+            client.borrow_readonly_connection(client.clone())
+        } else {
+            client.borrow_connection(client.clone())
+        };
+
+        conn.and_then(|mut conn| {
+            conn.execute_command(
+                Some(self.graph.graph_name()),
+                command,
+                None,
+                Some(&[query.as_str(), "--compact"]),
+            )
+        })
     }
 }
 
@@ -440,10 +452,14 @@ impl<'a, Out> ProcedureQueryBuilder<'a, Out, AsyncGraph> {
             generate_procedure_call(self.procedure_name, self.args, self.yields);
         let query = construct_query(query_string, params.as_ref());
 
-        self.graph
-            .get_client()
-            .borrow_connection(self.graph.get_client().clone())
-            .await?
+        let client = self.graph.get_client();
+        let conn = if self.readonly {
+            client.borrow_readonly_connection(client.clone()).await
+        } else {
+            client.borrow_connection(client.clone()).await
+        };
+
+        conn?
             .execute_command(
                 Some(self.graph.graph_name()),
                 command,